+++ src/oersted/__init__.py
@@ -1,5 +1,8 @@
 """Python bindings for oersted"""
 
+from .materials import MU0, FreeSpace, LinearMaterial, NonlinearMaterial, BHCurve
+from .solver import Solver, DirectSolver, OctreeSolver
+
 from .biotsavart import (
     bfield_direct,
     bfield_octree,
@@ -10,9 +13,10 @@ from .biotsavart import (
     hfield_dipole,
     hfield_dipole_tetrahedrons,
 )
-from .materials import MU0, FreeSpace, LinearMaterial, NonlinearMaterial, BHCurve
+
 from .mesh import Mesh, mesh_step
 from .magnetization import mag_force
+
 from . import testing
 
 
@@ -34,4 +38,7 @@ __all__ = [
     "hfield_dipole",
     "hfield_dipole_tetrahedrons",
     "mag_force",
+    "Solver",
+    "DirectSolver",
+    "OctreeSolver",
 ]