+++ src/oersted/biotsavart.py
@@ -16,41 +16,10 @@ from ._oersted import (
 )
 
 from .mesh import Mesh, CentroidMesh
+from .solver import DirectSolver, OctreeSolver
 
 # For typing; currently unused
 Nx3Array = NDArray[float64]
-
-
-class Solver:
-    _n_threads: int
-
-    @property
-    def n_threads(self):
-        return self._n_threads
-
-
-class DirectSolver(Solver):
-    def __init__(self, n_threads: int = 0):
-        self._n_threads = n_threads
-
-
-class OctreeSolver(Solver):
-    _theta: float
-    _leaf_threshold: int
-
-    def __init__(self, theta: float = 0.25, leaf_threshold: int = 16, n_threads: int = 0):
-
-        self._theta = theta
-        self._leaf_threshold = leaf_threshold
-        self._n_threads = n_threads
-
-    @property
-    def theta(self):
-        return self._theta
-
-    @property
-    def leaf_threshold(self):
-        return self._leaf_threshold
 
 
 def b_field(source: Mesh | CentroidMesh, targets: NDArray[float64], solver: DirectSolver | OctreeSolver | None = None) -> NDArray[float64]: